Move filter data to constant data
