+++ packages/frontend/src/constants/Data.ts
@@ -1,0 +1,3 @@
+export const FileFormats = ['PDF', 'Excel', 'Image', 'PowerPoint', 'Word', 'PlainText', 'Muut'];
+
+export const FinnishRegions = ['Etelä', 'Länsi', 'Itä', 'Pohjoinen'];
